requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-groq
langchain-core
langchain-pinecone
langchain-huggingface
pinecone
pypdf
python-dotenv
sentence-transformers
langchain-classic